# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Client Onboarding Platform API
  version: 0.4.0-phase-3
  description: Identity, tenant-scoped client/project core, and versioned workflow engine.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/organizations/{organizationId}:
    parameters:
      - name: organizationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Organizations]
      security: [{ cookieAuth: [] }]
      summary: Read the authenticated tenant only
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Organizations]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, version]
              properties:
                name: { type: string, maxLength: 160 }
                version: { type: integer, format: int64, minimum: 0 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/permissions:
    get:
      tags: [RBAC]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/roles:
    get:
      tags: [RBAC]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [RBAC]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleRequest' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /api/v1/roles/{roleId}:
    patch:
      tags: [RBAC]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
        - name: roleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/roles/{roleId}/archive:
    post:
      tags: [RBAC]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
        - name: roleId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: version
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/organization-members:
    get:
      tags: [Organizations]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Organizations]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, displayName, roleId]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                displayName: { type: string, maxLength: 160 }
                roleId: { type: string, format: uuid }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /api/v1/organization-members/{membershipId}:
    patch:
      tags: [Organizations]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
        - name: membershipId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleId, status, version]
              properties:
                roleId: { type: string, format: uuid }
                status: { type: string, enum: [ACTIVE, SUSPENDED, ARCHIVED] }
                version: { type: integer, format: int64, minimum: 0 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: BOS_SESSION
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
  responses:
    Success:
      description: Standard success envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    RoleRequest:
      type: object
      required: [name, permissions, version]
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 240 }
        permissions:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 80 }
        version: { type: integer, format: int64, minimum: 0 }
...
    SuccessEnvelope:
      type: object
      required: [success, data, meta, requestId]
      properties:
        success: { const: true }
        data: {}
        meta: { type: object, additionalProperties: true }
        requestId: { type: string, format: uuid }
    ErrorEnvelope:
      type: object
      required: [success, error, requestId]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, fieldErrors]
          properties:
            code: { type: string }
            message: { type: string }
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
        requestId: { type: string, format: uuid }